d3 diagram added and shows language
The information doesn't show until you click the button. A function in script.js shows the value from data in a diagram. The languages is in the same order that the values are in the diagram. All repositories and their languages can be printed if the #profile isn't commented out.